+++ tools/docker/tuleap-aio/mysql-utils.sh
@@ -6,6 +6,9 @@ set +x
 
 # Starts the DB and upgrade the data
 start_mysql() {
+    if [ -n "$DB_HOST" ]; then
+        return;
+    fi
     # old password must be disabled for php 5.6 / mysqlnd
     sed -i -e 's/^old_passwords\(.*\)/#old_passwords\1/' /etc/my.cnf
     codendiadm_pass=$(./interpolate_tuleap_var.php /etc/tuleap/conf/database.inc sys_dbpasswd)
@@ -13,10 +16,7 @@ start_mysql() {
     echo "Start mysql"
     /usr/bin/python /usr/lib/python2.6/site-packages/supervisor/pidproxy.py /var/run/mysqld/mysqld.pid /usr/bin/mysqld_safe &
     sleep 1
-    while ! mysql -ucodendiadm -p$codendiadm_pass -e "show databases" >/dev/null; do
-	echo "Wait for the db"
-	sleep 1
-    done
+    wait_for_db localhost codendiadm $codendiadm_pass
 
     # Update password when switching from old_password
     if grep -q '#old_passwords' /etc/my.cnf; then
@@ -24,8 +24,21 @@ start_mysql() {
     fi
 }
 
+wait_for_db() {
+    host=$1
+    user=$2
+    password=$3
+    while ! mysql -h$host -u$user -p$password -e "show databases" >/dev/null; do
+        echo "Wait for the db"
+        sleep 1
+    done
+}
+
 # Stop Mysql
 stop_mysql() {
+    if [ -n "$DB_HOST" ]; then
+        return;
+    fi
     echo "Stop mysql"
     PID=$(cat /var/run/mysqld/mysqld.pid)
     kill -15 $PID
